deduplicate Http4sBackend across JVM and Native
Extract the full backend implementation into a shared abstract class
Http4sBackendBase plus a private companion holding the platform-agnostic
factory helpers (defaultCompressionHandlers, usingClient, stub). Each
platform keeps a thin class wrapper and an object Http4sBackend with the
platform-specific factories (Blaze + Ember on JVM; Ember with Network on
Native), delegating the shared bits to Http4sBackendBase.

A previous attempt mixed a shared platform trait into object Http4sBackend.
That made the dotc reader resolve every method in the trait — including
the Blaze-typed ones — whenever any member of Http4sBackend was used,
which broke mdoc on classpaths without blaze-client. Keeping the methods
inline in each platform object sidesteps this.
